Tweaked Gravity, Removed Warnings

+++ Grid.cpp
@@ -1,20 +1,18 @@
 #include "Grid.h"
 
-void Grid::init(float xSize, float ySize){
+void Grid::init(float xSize, float ySize)
+{
     grid.resize(xSize);
     
-    for (int i = 0; i < grid.size(); i++) {
+    for (int i = 0; i < (int)grid.size(); i++)
+    {
         grid[i].resize(ySize);
     }
     //srand (time(NULL)); // Turn this off if you want a random value uniform across all simulations.
 
     int x, y;
-	//float r, g, b;
-	//float forceX, forceY;
-	//forceX = 0.0;//0.064;
 	for(x = 0; x < xSize; x++)
 	{
-		//forceY = 0.0f;
 		for(y = 0; y < ySize; y++)
 		{
 			Node a;
